+++ test/api/v3/integration/user/stats/POST-user_allocate.test.js
@@ -1,7 +1,7 @@
 import {
   generateUser,
   translate as t,
-} from '../../../../helpers/api-integration/v3';
+} from '../../../../../helpers/api-integration/v3';
 
 describe('POST /user/allocate', () => {
   let user;
